feat: simplify settings IA — 4 tabs, primary "restore latest" backup CTA
Settings tab IA: 5 → 4

| Before                                                         | After                                  |
|----------------------------------------------------------------|----------------------------------------|
| Nodes / Backups / Notifications / Updates / Advanced           | Nodes / System / Integrations / Backups |

Advanced was a kitchen-sink (server identity, log level, template
variables, registries) and Notifications had its own tab for one
section (SMTP). Both reorganized into purpose-built groups:

- System  →  Server Identity + Updates + Log Level
- Integrations →  Email (SMTP) + Template Registries + Template Variables

Each former section is now a standalone component under
`_lib/sections/` so the new pages are thin compositions:

- `_lib/sections/ServerIdentitySection.tsx`
- `_lib/sections/UpdatesSection.tsx`
- `_lib/sections/EmailNotificationsSection.tsx`
- `_lib/sections/TemplateRegistriesSection.tsx`
- `_lib/sections/TemplateVariablesSection.tsx`

Old routes deleted with no compatibility shim — there is no
production install yet so old `/settings/advanced`, `/notifications`,
`/updates` URLs simply 404.

Backups: primary "Restore latest snapshot" CTA

The selective-restore overlay (40+ toggles, 1500+ LOC) is the right
tool for the rare granular case but it dominated the page. Added a
single emerald CTA card at the very top of the backups page when at
least one backup exists:

  [↻] Restore latest snapshot
      One-click restore of <fileName> (<date>, <size>)
      Need granular control? Selective restore…
                                              [ Restore ]

Backed by the existing `/api/settings/backups/restore` endpoint with
no `selection` payload — it already supports the "restore everything"
path via `restoreSystemBackup(fileName)`. ConfirmModal in front of it
to avoid foot-gun. The old "Restore from Backup" header button is
relabeled "Selective restore…" so the two paths are clearly distinct.

Smoke-tested: GET /settings/{nodes,system,integrations,backups} → 200,
old /settings/{notifications,updates,advanced} → 404, no console
errors during page render in the dev container.
