+++ src/app/(dashboard)/settings/integrations/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import EmailNotificationsSection from '../_lib/sections/EmailNotificationsSection';
+import TemplateRegistriesSection from '../_lib/sections/TemplateRegistriesSection';
+import TemplateVariablesSection from '../_lib/sections/TemplateVariablesSection';
+
+export default function IntegrationsSettingsPage() {
+  return (
+    <>
+      <EmailNotificationsSection />
+      <TemplateRegistriesSection />
+      <TemplateVariablesSection />
+    </>
+  );
+}
